uos: Add BUILD_ID to os.uname()
It'll be convenient to store build ID in device firmware to help track
builds more effectively with building Micropython in CI/CD environments
such as Zuul or Jenkins.

This commit adds support for including the BUILD_ID in the output of
os.uname(). BUILD_ID is retrieved from an environment variable during
the build process, and if not set, defaults to "<no build-id>".
MICROPY_BUILD_ID is defined in makeversionhdr.py and used everywhere
with MICROPY_BUILD_DATE.

Change-Id: I48b418e1be9f6640d1274b2c99e9b650808bd26f